+++ printing/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add PdfPreview Widget
 - Implement Printing.raster() on Flutter Web
+- Fix Swift 5 deprecated function
 
 ## 3.3.1